refactor/Type Safety 네비게이션 적용
