+++ app/src/main/java/org/sopt/at/views/search/SearchScreen.kt
@@ -7,7 +7,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
-import org.sopt.at.views.navigation.Route
+import org.sopt.at.views.navigation.Screen
 
 @Composable
 fun SearchScreen(
@@ -19,7 +19,7 @@ fun SearchScreen(
         contentAlignment = Alignment.Center
     ) {
         Text(
-            text = Route.SEARCH,
+            text = Screen.Search.route,
             fontWeight = FontWeight.Bold
         )
     }
