+++ app/src/main/java/org/sopt/at/views/shorts/ShortsScreen.kt
@@ -7,7 +7,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
-import org.sopt.at.views.navigation.Route
+import org.sopt.at.views.navigation.Screen
 
 @Composable
 fun ShortsScreen(
@@ -19,7 +19,7 @@ fun ShortsScreen(
         contentAlignment = Alignment.Center
     ) {
         Text(
-            text = Route.SHORTS,
+            text = Screen.Shorts.route,
             fontWeight = FontWeight.Bold
         )
     }
